+++ src/smartcontracts/typesystem/generic.ts
@@ -1,6 +1,8 @@
 import { guardValueIsSet } from "../../utils";
+import { CollectionOfTypedValues } from "./collections";
 import { Type, TypedValue, NullType, TypePlaceholder } from "./types";
 
+// TODO: Move to a new file, "genericOption.ts"
 export class OptionType extends Type {
     constructor(typeParameter: Type) {
         super("Option", [typeParameter]);
@@ -17,12 +19,14 @@ export class OptionType extends Type {
     }
 }
 
+// TODO: Move to a new file, "genericList.ts"
 export class ListType extends Type {
     constructor(typeParameter: Type) {
         super("List", [typeParameter]);
     }
 }
 
+// TODO: Move to a new file, "genericOption.ts"
 export class OptionValue extends TypedValue {
     private readonly value: TypedValue | null;
 
@@ -72,10 +76,11 @@ export class OptionValue extends TypedValue {
     }
 }
 
+// TODO: Move to a new file, "genericList.ts"
 // TODO: Rename to ListValue, for consistency (though the term is slighly unfortunate).
 // Question for review: or not?
 export class List extends TypedValue {
-    private readonly items: TypedValue[];
+    private readonly backingCollection: CollectionOfTypedValues;
 
     /**
      *
@@ -87,7 +92,7 @@ export class List extends TypedValue {
 
         // TODO: assert items are of type type.getFirstTypeParameter()
 
-        this.items = items;
+        this.backingCollection = new CollectionOfTypedValues(items);
     }
 
     static fromItems(items: TypedValue[]): List {
@@ -96,35 +101,23 @@ export class List extends TypedValue {
         }
 
         let typeParameter = items[0].getType();
-        return new List(typeParameter, items);
+        let listType = new ListType(typeParameter);
+        return new List(listType, items);
     }
 
     getLength(): number {
-        return this.items.length;
+        return this.backingCollection.getLength();
     }
 
     getItems(): ReadonlyArray<TypedValue> {
-        return this.items;
+        return this.backingCollection.getItems();
     }
 
     valueOf(): any[] {
-        return this.items.map((item) => item.valueOf());
+        return this.backingCollection.valueOf();
     }
 
     equals(other: List): boolean {
-        if (this.getLength() != other.getLength()) {
-            return false;
-        }
-
-        for (let i = 0; i < this.getLength(); i++) {
-            let selfItem = this.items[i];
-            let otherItem = other.items[i];
-
-            if (!selfItem.equals(otherItem)) {
-                return false;
-            }
-        }
-
-        return true;
+        return this.backingCollection.equals(other.backingCollection);
     }
 }