+++ src/token.ts
@@ -12,8 +12,7 @@ export class Token {
     name: string = ''; // Token name (eg. MyTokenName123)
     type: TokenType = TokenType.Fungible;
     owner: Address = new Address();
-    minted: BigNumber = new BigNumber(0);
-    burnt: BigNumber = new BigNumber(0);
+    supply: string = '0'; // Circulating supply = initial minted supply + local mints - local burns
     decimals: number = 18;
     isPaused: boolean = false;
     canUpgrade: boolean = false;
@@ -33,12 +32,11 @@ export class Token {
     }
 
     static fromHttpResponse(response: {
-        token: string,
+        identifier: string,
         name: string,
         type: string,
         owner: string,
-        minted: string,
-        burnt: string,
+        supply: string,
         decimals: number,
         isPaused: boolean,
         canUpgrade: boolean,
@@ -50,12 +48,11 @@ export class Token {
         canWipe: boolean
     }): Token {
         return new Token({
-            identifier: response.token,
+            identifier: response.identifier,
             name: response.name,
             type: TokenType[response.type as keyof typeof TokenType],
             owner: new Address(response.owner),
-            minted: new BigNumber(response.minted),
-            burnt: new BigNumber(response.burnt),
+            supply: response.supply,
             decimals: response.decimals,
             isPaused: response.isPaused,
             canUpgrade: response.canUpgrade,
@@ -69,15 +66,14 @@ export class Token {
     }
 
     static fromTokenProperties(tokenIdentifier: string, results: any[]): Token {
-        let [tokenName, tokenType, owner, totalMinted, totalBurned, ...propertiesBuffers] = results;
+        let [tokenName, tokenType, owner, supply, ...propertiesBuffers] = results;
         let properties = parseTokenProperties(propertiesBuffers);
         return new Token({
             identifier: tokenIdentifier,
             type: TokenType[tokenType.toString() as keyof typeof TokenType],
             name: tokenName.toString(),
             owner,
-            minted: new BigNumber(totalMinted),
-            burnt: new BigNumber(totalBurned),
+            supply: supply,
             decimals: properties.NumDecimals.toNumber(),
             isPaused: properties.IsPaused,
             canUpgrade: properties.CanUpgrade,